refactor: extract SummaryResultHeader
