+++ src/app/(main)/summary/_components/summaryResultCard/index.ts
@@ -1,0 +1,1 @@
+export { SummaryResultHeader } from './SummaryResultHeader'
